Add python2 shebang to all Python scripts
These scripts are already executable due to their chmod +x, but they
are missing a shebang to tell the shell which interpreter is respons-
ible for executing them. The convention is to use the "env" command
to find an appropriate Python that may be in the user's environment
instead of hard coding a path, as the user might have a virtual env
loaded or might have Python 3 as the default "python" on the system.